Spent some time trying to get photologue to put things in order. The function to map to the proxy model from any methods that spit out the original model solved the problem.
A simpler implementation would be to alter photologues own manager methods to
include an order_by('date_added') on the filters. I've tried this and it works.

The this implementation we must keep the photos in the projects galleries sepparate
from those in the explore gallery otherwise photologue will display them in
the wrong url. May need to take off the View all galleries too.